+++ changelog.rst
@@ -12,6 +12,8 @@ and this project adheres to `Semantic Versioning <https://semver.org/spec/v2.0.0
 Added
 ^^^^^
 
+- ``zero_props`` and ``pole_props`` argument to ``plane_plots.*plane*``-functions
+  that allows more detailed styling of pole/zero markers.
 - ``multiplicity_props`` argument to ``plane_plots.*plane*``-functions that allows
   styling of pole/zero multiplicity texts.
 - This changelog.